+++ src/serviceconnector-passwordless/HISTORY.rst
@@ -2,6 +2,14 @@
 
 Release History
 ===============
+2.0.1
+++++++
+* Fix old api-version issue
+
+2.0.0
+++++++
+* Update to be compatible with Azure CLI 2.60.0
+
 1.0.3
 ++++++
 * Fix no attribute error